Added tests & template for Inaktivera utskick #57
