+++ html/utskick.html
@@ -69,7 +69,7 @@
 <button 
     class="btn btn-primary"
     v-if="edit=='' && me.admin"
-    v-on:click="edit={'type':'utskick','data':{'id':'','datum':'','typ':'e-post','rubrik':'','text':'','kodstugor_id':''}}"
+    v-on:click="edit={'type':'utskick','data':{'id':'','datum':'','status':'aktiv','typ':'e-post','rubrik':'','text':'','kodstugor_id':''}}"
     >
     Skapa ny
 </button>
@@ -98,6 +98,14 @@
                 <select name="typ" required v-model:value="edit.data.typ" class="form-control">
                     <option value="e-post">E-Post</option>
                     <option value="sms">SMS</option>
+                </select>
+            </div>
+            <div class="form-group">
+                <label>Status</label>
+                <select name="status" required v-model:value="edit.data.status" class="form-control">
+                    <option value="aktiv">Aktiv</option>
+                    <option value="inaktiv">Inaktiv</option>
+                    <option value="skickad">Skickad</option>
                 </select>
             </div>
             <div class="form-group" v-if="edit.data.typ=='e-post'">
@@ -129,7 +137,7 @@
             </div>
             <div class="form-group">
                 <button v-if="edit.data.id==''" type="submit" class="btn btn-primary">Spara som ny</button>
-                <button v-if="edit.data.status=='väntar'&&edit.data.id!==''" type="submit" class="btn btn-primary">Spara</button>
+                <button v-if="edit.data.status!=='skickad'&&edit.data.id!==''" type="submit" class="btn btn-primary">Spara</button>
                 <button v-on:click="radera_utskick(edit.data.id)" type="button" class="btn btn-danger" v-if="me.admin && edit.data.id!==''" >Radera</button>
                 <button v-on:click="edit=''" type="button" class="btn btn-secondary">Avbryt</button>
                 <button v-on:click="edit.data.id=''" v-if="edit.data.id!==''" type="button" class="btn btn-secondary">Skapa ett nytt utskick med denna som mall</button>
