+++ managedata/invite.py
@@ -60,53 +60,57 @@ def reply(request):
     return static_file('static/reply_done.html')
 
 
+def send_invites_once():
+    found = db.cursor.execute('''
+        SELECT
+            kodstugor.namn AS kodstuga,
+            kodstugor.typ AS kodstuga_typ,
+            deltagare.id AS deltagare_id,
+            deltagare.status AS deltagare_status,
+            deltagare.fornamn AS deltagare_fornamn,
+            kontaktpersoner.epost AS kontaktperson_epost
+        FROM
+            deltagare
+        INNER JOIN kontaktpersoner_deltagare
+            ON deltagare.id=kontaktpersoner_deltagare.deltagare_id
+        INNER JOIN kontaktpersoner
+           ON kontaktpersoner.id=kontaktpersoner_deltagare.kontaktpersoner_id
+        INNER JOIN kodstugor
+           ON deltagare.kodstugor_id=kodstugor.id
+        WHERE deltagare.status = "inbjudan"
+        ORDER BY deltagare.id;
+        ''')
+
+    def to_headers(row):
+        ut = {}
+        for idx, col in enumerate(found.description):
+            ut[col[0]] = row[idx]
+        return ut
+    deltagare_id = False
+    for row in list(map(to_headers, found.fetchall())):
+        if not deltagare_id:
+            deltagare_id = row["deltagare_id"]
+        if row["deltagare_id"] != deltagare_id:
+            break
+        message = texter.get_one("Erbjudande om plats " + row["kodstuga_typ"])["text"]
+        link = "https://besk.kodcentrum.se/reply?id=" + row["deltagare_id"]
+        message = message.replace(
+            "%namn%", row["deltagare_fornamn"]).replace(
+            "%kodstuga%", row["kodstuga"]).replace(
+            "%länk%", link)
+        send_email(row["kontaktperson_epost"], "Erbjudande om plats", message)
+    db.cursor.execute('''
+            UPDATE
+                deltagare
+            SET
+                status = "inbjuden"
+            WHERE
+                id = ?;
+    ''', (deltagare_id,))
+    db.commit()
+
+
 def send_invites():
     while True:
         sleep(2)
-        found = db.cursor.execute('''
-            SELECT
-                kodstugor.namn AS kodstuga,
-                kodstugor.typ AS kodstuga_typ,
-                deltagare.id AS deltagare_id,
-                deltagare.status AS deltagare_status,
-                deltagare.fornamn AS deltagare_fornamn,
-                kontaktpersoner.epost AS kontaktperson_epost
-            FROM
-                deltagare
-            INNER JOIN kontaktpersoner_deltagare
-                ON deltagare.id=kontaktpersoner_deltagare.deltagare_id
-            INNER JOIN kontaktpersoner
-               ON kontaktpersoner.id=kontaktpersoner_deltagare.kontaktpersoner_id
-            INNER JOIN kodstugor
-               ON deltagare.kodstugor_id=kodstugor.id
-            WHERE deltagare.status = "inbjudan"
-            ORDER BY deltagare.id;
-            ''')
-
-        def to_headers(row):
-            ut = {}
-            for idx, col in enumerate(found.description):
-                ut[col[0]] = row[idx]
-            return ut
-        deltagare_id = False
-        for row in list(map(to_headers, found.fetchall())):
-            if not deltagare_id:
-                deltagare_id = row["deltagare_id"]
-            if row["deltagare_id"] != deltagare_id:
-                break
-            message = texter.get_one("Erbjudande om plats " + row["kodstuga_typ"])["text"]
-            link = "https://besk.kodcentrum.se/reply?id=" + row["deltagare_id"]
-            message = message.replace(
-                "%namn%", row["deltagare_fornamn"]).replace(
-                "%kodstuga%", row["kodstuga"]).replace(
-                "%länk%", link)
-            send_email(row["kontaktperson_epost"], "Erbjudande om plats", message)
-        db.cursor.execute('''
-                UPDATE
-                    deltagare
-                SET
-                    status = "inbjuden"
-                WHERE
-                    id = ?;
-        ''', (deltagare_id,))
-        db.commit()
+        send_invites_once()
